+++ CHANGELOG.md
@@ -6,6 +6,10 @@ project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Added per-series `AreaFillBrush` support for multi-series `LineChart` charts.
+
 ## [0.3.0] - 2026-07-22
 
 ### Added